// src/generated/model/navEnhetFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface NavEnhetFrontend {
    orgnr?: string;
    enhetsnr?: string;
    enhetsnavn: string;
    kommunenavn?: string;
    kommuneNr?: string;
    behandlingsansvarlig?: string;
    valgt?: boolean;
    isMottakMidlertidigDeaktivert?: boolean;
    isMottakDeaktivert?: boolean;
}